Made a start on game database
Appears things have changed drastically since my ventures with forum app
some of the classes I used are now deprecated. These classes still work
for this now but would be good to learn the new way to approach cursors
in a database *reminder Callum to look at in spare time *. For the
purpose of this fun game should not matter.